requirements.txt
pytest
black
transformers
flask
torch
python-dotenv